api_uat/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging & Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.3
httpx==0.26.0

# Production Server
gunicorn==21.2.0
